update test - separating test data and minor update
