+++ test/test.html
@@ -18,7 +18,8 @@
   <!-- Intergallactic library -->
   <script src="../dist/intergallactic.min.js"></script>
   <!-- load all test data -->
-  <script src="utils/testData.js"></script>
+  <script src="utils/common.td.js"></script>
+  <script src="utils/conversion.td.js"></script>
 
   <!-- test files -->
   <script src="intergallactic.test.js"></script>
